+++ nannou/src/lib.rs
@@ -13,9 +13,6 @@
 //! examples](https://github.com/nannou-org/nannou/tree/master/examples) to get an idea of how
 //! nannou applications are structured and how the API works.
 
-pub use conrod_core;
-pub use conrod_wgpu;
-pub use conrod_winit;
 pub use daggy;
 pub use find_folder;
 pub use lyon;
@@ -25,7 +22,6 @@ pub use self::app::{App, LoopMode};
 pub use self::draw::Draw;
 pub use self::event::Event;
 pub use self::frame::Frame;
-pub use self::ui::Ui;
 #[doc(inline)]
 pub use nannou_core::{color, glam, math, rand};
 #[doc(inline)]
@@ -46,7 +42,6 @@ pub mod prelude;
 pub mod state;
 pub mod text;
 pub mod time;
-pub mod ui;
 pub mod window;
 
 /// Begin building the `App`.